[FIX] Fix compilation error and warnings in GCC and Clang.

+++ Source/Core/Public/Atom/Hash/Private/T1HashGenerator.h
@@ -90,7 +90,7 @@ namespace Atom::Private
             for (usize processed = 0; processed < dataSize; processed += maxInput)
             {
                 _impl.Update((byte*)data + processed,
-                    Math::Min(maxInput, dataSize - processed));
+                    (uint32_t)Math::Min(maxInput, dataSize - processed));
             }
 
             return *this;
